[Refactoring] Refactor package and classes names
