+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.BlogViewer">
-        <activity android:name=".authors.MainActivity">
+        <activity android:name="com.paulacr.blogviewer.feature.authors.AuthorsListActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -20,7 +20,7 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".details.DetailsActivity"/>
+        <activity android:name="com.paulacr.blogviewer.feature.posts.PostsActivity"/>
     </application>
 
 </manifest>
