+++ app/src/main/java/com/paulacr/blogviewer/di/ApplicationComponent.kt
@@ -1,7 +1,7 @@
 package com.paulacr.blogviewer.di
 
 import android.app.Application
-import com.paulacr.blogviewer.authors.MainActivity
+import com.paulacr.blogviewer.feature.authors.AuthorsListActivity
 import com.paulacr.data.network.NetworkModule
 import com.paulacr.data.repository.DataModule
 import dagger.BindsInstance
@@ -10,7 +10,7 @@ import dagger.Component
 @Component(modules = [NetworkModule::class, DataModule::class])
 interface ApplicationComponent {
 
-    fun inject(activity: MainActivity)
+    fun inject(activity: AuthorsListActivity)
 
     @Component.Builder
     interface Builder {
